# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/organizations/{organizationId}/workflows/types": {
      "get": {
        "operationId": "WorkflowsController_getTypes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Workflows"
        ]
      }
    },
    "/api/organizations/{organizationId}/workflows": {
      "post": {
        "operationId": "WorkflowsController_create",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Workflows"
        ]
      },
      "get": {
        "operationId": "WorkflowsController_findAll",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Workflows"
        ]
      }
    },
    "/api/organizations/{organizationId}/workflows/{workflowId}": {
      "get": {
        "operationId": "WorkflowsController_findOne",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Workflows"
        ]
      },
      "patch": {
        "operationId": "WorkflowsController_update",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Workflows"
        ]
      },
      "delete": {
        "operationId": "WorkflowsController_remove",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Workflows"
        ]
      }
    },
    "/api/organizations/{organizationId}/workflows/{workflowId}/activate": {
      "post": {
        "operationId": "WorkflowsController_activate",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Workflows"
        ]
      }
    },
    "/api/organizations/{organizationId}/workflows/{workflowId}/deactivate": {
      "post": {
        "operationId": "WorkflowsController_deactivate",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Workflows"
        ]
      }
    },
    "/api/organizations/{organizationId}/workflows/{workflowId}/test": {
      "post": {
        "operationId": "WorkflowsController_test",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Workflows"
        ]
      }
    },
    "/api/organizations/{organizationId}/workflows/{workflowId}/runs": {
      "get": {
        "operationId": "WorkflowsController_getWorkflowRuns",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Workflows"
        ]
      }
    },
...
    "/api/organizations/{organizationId}/uploads/image": {
      "post": {
        "operationId": "UploadsController_uploadImage",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Uploads"
        ]
      }
    },
    "/api/organizations/{organizationId}/uploads/{filename}": {
      "delete": {
        "operationId": "UploadsController_deleteImage",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Uploads"
        ]
      }
    },
    "/api/organizations/{organizationId}/uploads/{category}/{filename}": {
      "get": {
        "operationId": "UploadsController_getImage",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Uploads"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenEOS API",
    "description": "\n# OpenEOS API Documentation\n\nOpen Event Ordering System - REST API für Kassensysteme bei Vereinen und Veranstaltungen.\n\n## Authentifizierung\n\nDie API verwendet JWT Bearer Tokens für die Authentifizierung. Nach dem Login erhältst du einen Access Token (30 Min gültig) und einen Refresh Token (7 Tage, als httpOnly Cookie).\n\n### Endpoints ohne Authentifizierung\n- `POST /api/auth/register` - Registrierung\n- `POST /api/auth/login` - Login\n- `POST /api/auth/refresh` - Token erneuern\n- `POST /api/auth/forgot-password` - Passwort vergessen\n- `POST /api/auth/reset-password` - Passwort zurücksetzen\n- `GET /api/health/*` - Health Checks\n- `GET /api/invitations/:token` - Einladung abrufen\n- `/api/public/order/*` - Online-Bestellungen (mit Session-Token)\n\n## Multi-Tenant\n\nDie API ist multi-tenant-fähig. Die meisten Endpoints erfordern eine `organizationId` im Pfad.\n\n## Rollen\n\n- **ADMIN** - Vollzugriff auf Organisation\n- **MANAGER** - Verwaltung von Produkten, Events, Bestellungen\n- **CASHIER** - Kassen-Operationen\n- **KITCHEN** - Küchen-Display\n- **DELIVERY** - Ausgabe-Display\n\n## WebSocket\n\nReal-time Updates über Socket.io unter `/` mit Namespaces:\n- `/pos` - Kassen-Clients\n- `/display` - Küchen/Ausgabe-Displays\n- `/printer` - Drucker-Agenten\n- `/admin` - Admin-Dashboard\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "OpenEOS",
...
    },
    "license": {
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "X-Organization-Id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Organization-Id",
        "description": "Organization ID for multi-tenant requests"
      },
      "X-Session-Token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Token",
        "description": "Session token for online orders"
      }
    },
    "schemas": {
...
      "CreateWorkflowDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Küchenbonausgabe",
            "description": "Name des Workflows"
          },
          "description": {
            "type": "string",
            "example": "Automatische Bonausgabe für Küchenprodukte",
            "description": "Beschreibung des Workflows"
          },
          "triggerType": {
            "type": "string",
            "example": "order_created",
            "description": "Auslöser-Typ des Workflows"
          },
          "triggerConfig": {
            "type": "object",
            "example": {
              "categories": [
                "kitchen"
              ],
              "status": "confirmed"
            },
            "description": "Konfiguration des Auslösers"
          },
          "nodes": {
            "example": [
              {
                "id": "node1",
                "type": "print",
                "config": {}
              }
            ],
            "description": "Workflow-Knoten",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "edges": {
            "example": [
              {
                "source": "node1",
                "target": "node2"
              }
            ],
            "description": "Verbindungen zwischen Knoten",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Gibt an, ob der Workflow aktiv ist"
          }
        },
        "required": [
          "name",
          "triggerType"
        ]
      },
      "UpdateWorkflowDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Küchenbonausgabe",
            "description": "Name des Workflows"
          },
          "description": {
            "type": "string",
            "example": "Automatische Bonausgabe für Küchenprodukte",
            "description": "Beschreibung des Workflows"
          },
          "triggerConfig": {
            "type": "object",
            "example": {
              "categories": [
                "kitchen"
              ],
              "status": "confirmed"
            },
            "description": "Konfiguration des Auslösers"
          },
          "nodes": {
            "example": [
              {
                "id": "node1",
                "type": "print",
                "config": {}
              }
            ],
            "description": "Workflow-Knoten",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "edges": {
            "example": [
              {
                "source": "node1",
                "target": "node2"
              }
            ],
            "description": "Verbindungen zwischen Knoten",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Gibt an, ob der Workflow aktiv ist"
          }
        }
      },
      "TestWorkflowDto": {
        "type": "object",
        "properties": {
          "testData": {
            "type": "object",
            "example": {
              "orderId": "550e8400-e29b-41d4-a716-446655440000",
              "orderNumber": "B-001"
            },
            "description": "Testdaten für den Workflow"
          }
        }
      },
...
    }
  }
}